requirements.txt
flask>=3.0
gunicorn>=21.0
fast-flights>=2.2
typing_extensions>=4.0
playwright>=1.44
python-dotenv>=1.0
requests>=2.31
stripe>=8.0
